Update to main

+++ main script/custom_creator/server/server_event.lua
@@ -38,69 +38,107 @@ CreateServerCallback('custom_creator:server:get_list', function(source, callback
 	local playerId = tonumber(source)
 	local identifier_license = GetPlayerIdentifierByType(playerId, 'license')
 	local result = {
-		["published-races"] = {},
-		["saved-races"] = {}
+		[1] = {
+			class = "published-races",
+			data = {}
+		},
+		[2] = {
+			class = "saved-races",
+			data = {}
+		}
 	}
 	local template = {}
+	local isAdmin = false
+	local result_admin = {}
 	if identifier_license then
 		local identifier = identifier_license:gsub('license:', '')
+		local query = MySQL.query.await("SELECT `group`, race_creator FROM custom_race_users WHERE license = ?", {identifier})
+		if query and query[1] then
+			isAdmin = query[1].group == "admin"
+			template = json.decode(query[1].race_creator) or {}
+		end
+		local count = 0
 		for k, v in pairs(MySQL.query.await("SELECT * FROM custom_race_list")) do
 			if identifier == v.license then
 				local data = {
 					name = v.route_file:match("([^/]+)%.json$"),
-					img = v.route_image
+					img = v.route_image,
+					raceid = v.raceid
 				}
-				if v.published == "√" then
-					table.insert(result["published-races"], data)
+				if v.published ~= "x" then
+					table.insert(result[1].data, data)
 				else
-					table.insert(result["saved-races"], data)
+					table.insert(result[2].data, data)
 				end
 			end
+			if isAdmin then
+				if count > 500 then
+					count = 0
+					Citizen.Wait(0)
+				end
+				if not result_admin[v.category] then
+					result_admin[v.category] = {}
+				end
+				count = count + 1
+				table.insert(result_admin[v.category], {
+					name = v.route_file:match("([^/]+)%.json$"),
+					img = v.route_image,
+					raceid = v.raceid
+				})
+			end
 		end
-		if #result["published-races"] >= 2 then
-			table.sort(result["published-races"], function(a, b)
-				return string.lower(a.name) < string.lower(b.name)
+		local sorted_categories = {}
+		for category in pairs(result_admin) do
+			table.insert(sorted_categories, category)
+		end
+		if #sorted_categories >= 2 then
+			table.sort(sorted_categories, function(a, b)
+				return string.lower(a) < string.lower(b)
 			end)
 		end
-		if #result["saved-races"] >= 2 then
-			table.sort(result["saved-races"], function(a, b)
-				return string.lower(a.name) < string.lower(b.name)
-			end)
+		for i = 1, #sorted_categories do
+			result[2 + i] = {
+				class = sorted_categories[i],
+				data = result_admin[sorted_categories[i]]
+			}
 		end
-		local query = MySQL.query.await("SELECT race_creator FROM custom_race_users WHERE license = ?", {identifier})
-		if query and query[1] then
-			template = json.decode(query[1].race_creator) or {}
+		for i = 1, #result do
+			if #result[i].data >= 2 then
+				table.sort(result[i].data, function(a, b)
+					return string.lower(a.name) < string.lower(b.name)
+				end)
+			end
 		end
 	end
 	callback(result, template)
 end)
 
-CreateServerCallback('custom_creator:server:get_json', function(source, callback, title)
+CreateServerCallback('custom_creator:server:get_json', function(source, callback, id)
 	local playerId = tonumber(source)
-	local path, raceid, published, owner_name, thumbnail = nil, nil, nil, nil, nil
+	local path, raceid, published, category, thumbnail = nil, id, nil, nil, nil
 
 	local identifier_license = GetPlayerIdentifierByType(playerId, 'license')
 	if identifier_license then
 		local identifier = identifier_license:gsub('license:', '')
-		for k, v in pairs(MySQL.query.await("SELECT * FROM custom_race_list")) do
-			if v.route_file:match("([^/]+)%.json$") == title and identifier == v.license then
-				path = v.route_file
-				raceid = v.raceid
-				published = v.published == "√"
-				owner_name = GetPlayerName(playerId)
-				thumbnail = v.route_image
-				break
-			end
+		local query = MySQL.query.await("SELECT route_file, route_image, category, published FROM custom_race_list WHERE raceid = ?", {raceid})
+		if query and query[1] then
+			path = query[1].route_file
+			thumbnail = query[1].route_image
+			category = query[1].category
+			published = query[1].published ~= "x"
 		end
+
 	end
 	if path then
-		local data = json.decode(LoadResourceFile(GetCurrentResourceName(), path))
+		local data = json.decode(LoadResourceFile(string.find(string.lower(path), "custom_files") and GetCurrentResourceName() or "custom_races", path))
 		if data then
 			data.raceid = raceid
 			data.published = published
 			data.thumbnail = thumbnail
-			data.mission.gen.ownerid = owner_name
-			data.mission.gen.nm = title
+			if category ~= "Custom" then
+				data.mission.gen.ownerid = category
+			end
+			data.mission.gen.nm = path:match("([^/]+)%.json$")
 			callback(data)
 		else
 			callback(false)
@@ -246,24 +284,31 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 					CheckUserRole(discordId, function(bool)
 						if bool then
 							local found = false
-							local r_path = "/custom_files/" .. identifier
-							local a_path = GetResourcePath(resourceName) .. r_path
 							local path = nil
-							data.mission.gen.ownerid = GetPlayerName(playerId)
-							if not os.rename(a_path, a_path) then
-								-- Due to the sandboxing of lua, I am not sure whether os.rename and os.remove will be invalid in the future, so now I have rewritten it only for os.execute
-								-- Test on artifact 12911
-								-- More info: https://docs.fivem.net/docs/developers/sandbox/
-								CreateUserPath("mkdir \"" .. a_path .. "\"")
-							end
+							local og_license = nil
 							if data.raceid then
 								local sql_result = MySQL.query.await("SELECT * FROM custom_race_list WHERE raceid = @raceid", {['@raceid'] = data.raceid})
 								if sql_result and #sql_result > 0 then
-									path = sql_result[1].route_file
 									found = true
+									path = sql_result[1].route_file
+									og_license = sql_result[1].license
 								end
 							end
-							if path and (path:match("([^/]+)%.json$") ~= data.mission.gen.nm) then
+							local r_path = "/custom_files/" .. (og_license or identifier)
+							local a_path = GetResourcePath(resourceName) .. r_path
+							if not os.rename(a_path, a_path) then
+								if os and os.createdir then
+									os.createdir(a_path)
+								else
+									local success, _error = os.execute("mkdir \"" .. a_path .. "\"")
+									if not success or (_error == "Permission denied") then
+										action = "wrong-artifact"
+										print("Failed to save ^1" .. data.mission.gen.nm .. "^0. Please check if the ^2server artifact >= 13026 or <= 11895^0")
+										print("More info: https://docs.fivem.net/docs/developers/sandbox/")
+									end
+								end
+							end
+							if path and string.find(string.lower(path), "custom_files") and (path:match("([^/]+)%.json$") ~= data.mission.gen.nm) and (action ~= "wrong-artifact") then
 								os.remove(GetResourcePath(resourceName) .. path)
 							end
 							if action == "publish" then
@@ -281,32 +326,33 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 										if result then
 											data.raceid = result
 											data.published = true
+											data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 											SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
 											if GetResourceState("custom_races") == "started" then
 												TriggerEvent('custom_races:server:UpdateAllRace')
 											end
-											callback("success", result, GetPlayerName(playerId))
+											callback("success", result, data.mission.gen.ownerid)
 										else
 											callback(nil, nil, nil)
 										end
 									end)
 								else
-									MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ?, license = ? WHERE raceid = ?",
+									MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ? WHERE raceid = ?",
 									{
 										r_path .. "/" .. data.mission.gen.nm .. ".json",
 										data.thumbnail,
 										"√",
 										os.date("%Y/%m/%d %H:%M:%S", os.time()),
-										identifier,
 										data.raceid
 									}, function(result)
 										if result then
 											data.published = true
+											data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 											SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
 											if GetResourceState("custom_races") == "started" then
 												TriggerEvent('custom_races:server:UpdateAllRace')
 											end
-											callback("success", data.raceid, GetPlayerName(playerId))
+											callback("success", data.raceid, data.mission.gen.ownerid)
 										else
 											callback(nil, nil, nil)
 										end
@@ -314,21 +360,22 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 								end
 							elseif action == "update" then
 								if found then
-									MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ?, license = ? WHERE raceid = ?",
+									MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ? WHERE raceid = ?",
 									{
 										r_path .. "/" .. data.mission.gen.nm .. ".json",
 										data.thumbnail,
 										"√",
 										os.date("%Y/%m/%d %H:%M:%S", os.time()),
-										identifier,
 										data.raceid
 									}, function(result)
 										if result then
+											data.published = true
+											data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 											SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
 											if GetResourceState("custom_races") == "started" then
 												TriggerEvent('custom_races:server:UpdateAllRace')
 											end
-											callback("success", data.raceid, GetPlayerName(playerId))
+											callback("success", data.raceid, data.mission.gen.ownerid)
 										else
 											callback(nil, nil, nil)
 										end
@@ -352,31 +399,34 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 										if result then
 											data.raceid = result
 											data.published = false
+											data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 											SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
-											callback("success", result, GetPlayerName(playerId))
+											callback("success", result, data.mission.gen.ownerid)
 										else
 											callback(nil, nil, nil)
 										end
 									end)
 								else
-									MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ?, license = ? WHERE raceid = ?",
+									MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ? WHERE raceid = ?",
 									{
 										r_path .. "/" .. data.mission.gen.nm .. ".json",
 										data.thumbnail,
 										"x",
 										os.date("%Y/%m/%d %H:%M:%S", os.time()),
-										identifier,
 										data.raceid
 									}, function(result)
 										if result then
 											data.published = false
+											data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 											SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
-											callback("success", data.raceid, GetPlayerName(playerId))
+											callback("success", data.raceid, data.mission.gen.ownerid)
 										else
 											callback(nil, nil, nil)
 										end
 									end)
 								end
+							elseif action == "wrong-artifact" then
+								callback("wrong-artifact", nil, nil)
 							end
 						else
 							callback("denied", nil, nil)
@@ -395,24 +445,31 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 				end
 				if hasPermission then
 					local found = false
-					local r_path = "/custom_files/" .. identifier
-					local a_path = GetResourcePath(resourceName) .. r_path
 					local path = nil
-					data.mission.gen.ownerid = GetPlayerName(playerId)
-					if not os.rename(a_path, a_path) then
-						-- Due to the sandboxing of lua, I am not sure whether os.rename and os.remove will be invalid in the future, so now I have rewritten it only for os.execute
-						-- Test on artifact 12911
-						-- More info: https://docs.fivem.net/docs/developers/sandbox/
-						CreateUserPath("mkdir \"" .. a_path .. "\"")
-					end
+					local og_license = nil
 					if data.raceid then
 						local sql_result = MySQL.query.await("SELECT * FROM custom_race_list WHERE raceid = @raceid", {['@raceid'] = data.raceid})
 						if sql_result and #sql_result > 0 then
-							path = sql_result[1].route_file
 							found = true
+							path = sql_result[1].route_file
+							og_license = sql_result[1].license
 						end
 					end
-					if path and (path:match("([^/]+)%.json$") ~= data.mission.gen.nm) then
+					local r_path = "/custom_files/" .. (og_license or identifier)
+					local a_path = GetResourcePath(resourceName) .. r_path
+					if not os.rename(a_path, a_path) then
+						if os and os.createdir then
+							os.createdir(a_path)
+						else
+							local success, _error = os.execute("mkdir \"" .. a_path .. "\"")
+							if not success or (_error == "Permission denied") then
+								action = "wrong-artifact"
+								print("Failed to save ^1" .. data.mission.gen.nm .. "^0. Please check if the ^2server artifact >= 13026 or <= 11895^0")
+								print("More info: https://docs.fivem.net/docs/developers/sandbox/")
+							end
+						end
+					end
+					if path and string.find(string.lower(path), "custom_files") and (path:match("([^/]+)%.json$") ~= data.mission.gen.nm) and (action ~= "wrong-artifact") then
 						os.remove(GetResourcePath(resourceName) .. path)
 					end
 					if action == "publish" then
@@ -430,32 +487,33 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 								if result then
 									data.raceid = result
 									data.published = true
+									data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 									SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
 									if GetResourceState("custom_races") == "started" then
 										TriggerEvent('custom_races:server:UpdateAllRace')
 									end
-									callback("success", result, GetPlayerName(playerId))
+									callback("success", result, data.mission.gen.ownerid)
 								else
 									callback(nil, nil, nil)
 								end
 							end)
 						else
-							MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ?, license = ? WHERE raceid = ?",
+							MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ? WHERE raceid = ?",
 							{
 								r_path .. "/" .. data.mission.gen.nm .. ".json",
 								data.thumbnail,
 								"√",
 								os.date("%Y/%m/%d %H:%M:%S", os.time()),
-								identifier,
 								data.raceid
 							}, function(result)
 								if result then
 									data.published = true
+									data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 									SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
 									if GetResourceState("custom_races") == "started" then
 										TriggerEvent('custom_races:server:UpdateAllRace')
 									end
-									callback("success", data.raceid, GetPlayerName(playerId))
+									callback("success", data.raceid, data.mission.gen.ownerid)
 								else
 									callback(nil, nil, nil)
 								end
@@ -463,21 +521,22 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 						end
 					elseif action == "update" then
 						if found then
-							MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ?, license = ? WHERE raceid = ?",
+							MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ? WHERE raceid = ?",
 							{
 								r_path .. "/" .. data.mission.gen.nm .. ".json",
 								data.thumbnail,
 								"√",
 								os.date("%Y/%m/%d %H:%M:%S", os.time()),
-								identifier,
 								data.raceid
 							}, function(result)
 								if result then
+									data.published = true
+									data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 									SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
 									if GetResourceState("custom_races") == "started" then
 										TriggerEvent('custom_races:server:UpdateAllRace')
 									end
-									callback("success", data.raceid, GetPlayerName(playerId))
+									callback("success", data.raceid, data.mission.gen.ownerid)
 								else
 									callback(nil, nil, nil)
 								end
@@ -501,31 +560,34 @@ CreateServerCallback('custom_creator:server:save_file', function(source, callbac
 								if result then
 									data.raceid = result
 									data.published = false
+									data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 									SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
-									callback("success", result, GetPlayerName(playerId))
+									callback("success", result, data.mission.gen.ownerid)
 								else
 									callback(nil, nil, nil)
 								end
 							end)
 						else
-							MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ?, license = ? WHERE raceid = ?",
+							MySQL.update("UPDATE custom_race_list SET route_file = ?, route_image = ?, published = ?, updated_time = ? WHERE raceid = ?",
 							{
 								r_path .. "/" .. data.mission.gen.nm .. ".json",
 								data.thumbnail,
 								"x",
 								os.date("%Y/%m/%d %H:%M:%S", os.time()),
-								identifier,
 								data.raceid
 							}, function(result)
 								if result then
 									data.published = false
+									data.mission.gen.ownerid = data.mission.gen.ownerid or GetPlayerName(playerId)
 									SaveResourceFile(resourceName, r_path .. "/" .. data.mission.gen.nm .. ".json", json.encode(data), -1)
-									callback("success", data.raceid, GetPlayerName(playerId))
+									callback("success", data.raceid, data.mission.gen.ownerid)
 								else
 									callback(nil, nil, nil)
 								end
 							end)
 						end
+					elseif action == "wrong-artifact" then
+						callback("wrong-artifact", nil, nil)
 					end
 				else
 					callback("denied", nil, nil)
@@ -560,8 +622,10 @@ CreateServerCallback('custom_creator:server:cancel_publish', function(source, ca
 					path = result[1].route_file
 				end
 				local data = json.decode(LoadResourceFile(GetCurrentResourceName(), path))
-				data.published = false
-				SaveResourceFile(GetCurrentResourceName(), path, json.encode(data), -1)
+				if data then
+					data.published = false
+					SaveResourceFile(GetCurrentResourceName(), path, json.encode(data), -1)
+				end
 				callback(true)
 			else
 				callback(false)
